# rewear_api_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReWear API",
    "description": "Community Clothing Exchange Platform with Real-time Features",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/me/swaps": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Swaps",
        "description": "Get current user's swaps (sent and received)",
        "operationId": "get_user_swaps_api_v1_users_me_swaps_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "swap_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by 'sent' or 'received'",
              "title": "Swap Type"
            },
            "description": "Filter by 'sent' or 'received'"
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter by swap status",
              "title": "Status Filter"
            },
            "description": "Filter by swap status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "description": "Number of swaps to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of swaps to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of swaps to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of swaps to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SwapSummary"
                  },
                  "title": "Response Get User Swaps Api V1 Users Me Swaps Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/swaps/": {
      "post": {
        "tags": [
          "Swaps"
        ],
        "summary": "Create Swap Request",
        "description": "Create a new swap request with real-time notification",
        "operationId": "create_swap_request_api_v1_swaps__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwapCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Swaps"
        ],
        "summary": "List User Swaps",
        "description": "List current user's swaps (sent and received)",
        "operationId": "list_user_swaps_api_v1_swaps__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "swap_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by 'sent' or 'received'",
              "title": "Swap Type"
            },
            "description": "Filter by 'sent' or 'received'"
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by swap status",
              "title": "Status Filter"
            },
            "description": "Filter by swap status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SwapResponse"
                  },
                  "title": "Response List User Swaps Api V1 Swaps  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/swaps/{swap_id}/accept": {
      "put": {
        "tags": [
          "Swaps"
        ],
        "summary": "Accept Swap",
        "description": "Accept a swap request with real-time notification",
        "operationId": "accept_swap_api_v1_swaps__swap_id__accept_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "swap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Swap Id"
            }
          },
          {
            "name": "owner_response",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Response"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/swaps/{swap_id}/reject": {
      "put": {
        "tags": [
          "Swaps"
        ],
        "summary": "Reject Swap",
        "description": "Reject a swap request with real-time notification",
        "operationId": "reject_swap_api_v1_swaps__swap_id__reject_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "swap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Swap Id"
            }
          },
          {
            "name": "owner_response",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Response"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/swaps/{swap_id}/complete": {
      "put": {
        "tags": [
          "Swaps"
        ],
        "summary": "Complete Swap",
        "description": "Mark swap as completed with real-time notifications and points handling",
        "operationId": "complete_swap_api_v1_swaps__swap_id__complete_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "swap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Swap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/swaps/{swap_id}": {
      "get": {
        "tags": [
          "Swaps"
        ],
        "summary": "Get Swap",
        "description": "Get swap details (only if user is involved)",
        "operationId": "get_swap_api_v1_swaps__swap_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "swap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Swap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/swaps/{swap_id}/cancel": {
      "put": {
        "tags": [
          "Swaps"
        ],
        "summary": "Cancel Swap",
        "description": "Cancel a swap request (only by requester, only if pending)",
        "operationId": "cancel_swap_api_v1_swaps__swap_id__cancel_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "swap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Swap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/swaps/stats/summary": {
      "get": {
        "tags": [
          "Swaps"
        ],
        "summary": "Get Swap Stats",
        "description": "Get swap statistics for the current user",
        "operationId": "get_swap_stats_api_v1_swaps_stats_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Swap Stats Api V1 Swaps Stats Summary Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SwapCreate": {
        "properties": {
          "requester_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester Message"
          },
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "swap_type": {
            "type": "string",
            "title": "Swap Type"
          },
          "offered_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offered Item Id"
          },
          "points_offered": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points Offered"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "swap_type"
        ],
        "title": "SwapCreate",
        "description": "Schema for creating a swap request"
      },
...
      "SwapSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "swap_type": {
            "type": "string",
            "title": "Swap Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "points_offered": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points Offered"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "item": {
            "$ref": "#/components/schemas/app__schemas__swap__ItemSummary"
          },
          "offered_item": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/app__schemas__swap__ItemSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "swap_type",
          "status",
          "created_at",
          "item"
        ],
        "title": "SwapSummary",
        "description": "Brief swap summary for lists"
      },
...
      "app__schemas__swap__ItemSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "size": {
            "type": "string",
            "title": "Size"
          },
          "condition": {
            "type": "string",
            "title": "Condition"
          },
          "points_value": {
            "type": "integer",
            "title": "Points Value"
          },
          "primary_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Image Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "size",
          "condition",
          "points_value",
          "created_at"
        ],
        "title": "ItemSummary",
        "description": "Brief item summary for swaps"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}